+++ StarWarsDataAnalyzerWeb/Models/ErrorViewModel.cs
@@ -1,8 +1,9 @@
-namespace StarWarsDataAnalyzerWeb.Models;
-
-public class ErrorViewModel
+namespace StarWarsDataAnalyzerWeb.Models
 {
-    public string? RequestId { get; set; }
+    public class ErrorViewModel
+    {
+        public string RequestId { get; set; }
 
-    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    }
 }
